+++ pyproject.toml
@@ -92,8 +92,8 @@ include = ["pypsa"]
 
 [tool.pytest.ini_options]
 filterwarnings = [
-    "error::DeprecationWarning", # Raise all DeprecationWarnings as errors
-    "error::FutureWarning",      # Raise all FutureWarnings as errors
+    # "error::DeprecationWarning", # Raise all DeprecationWarnings as errors
+    # "error::FutureWarning",      # Raise all FutureWarnings as errors
     "ignore::matplotlib._api.deprecation.MatplotlibDeprecationWarning:pydev",
     # See https://github.com/microsoft/debugpy/issues/1623
 ]
