+++ pypsa/components/abstract.py
@@ -19,7 +19,8 @@ import pandas as pd
 import xarray
 from pyproj import CRS
 
-from pypsa.common import as_index, equals
+from pypsa.common import equals
+from pypsa.components.array import as_dynamic, as_xarray
 from pypsa.components.descriptors import (
     get_active_assets,
     get_activity_mask,
@@ -28,6 +29,7 @@ from pypsa.components.descriptors import (
     get_extendable_i,
     get_non_extendable_i,
 )
+from pypsa.components.transform import rename_component_names
 from pypsa.constants import DEFAULT_EPSG, DEFAULT_TIMESTAMP
 from pypsa.definitions.components import ComponentType
 from pypsa.definitions.structures import Dict
@@ -89,6 +91,10 @@ class Components(ComponentsData, ABC):
     # Methods
     # -------
 
+    # from pypsa.components.array
+    as_xarray = as_xarray
+    as_dynamic = as_dynamic
+
     # from pypsa.components.descriptors
     get_active_assets = get_active_assets
     get_activity_mask = get_activity_mask
@@ -96,6 +102,9 @@ class Components(ComponentsData, ABC):
     get_extendable_i = get_extendable_i
     get_non_extendable_i = get_non_extendable_i
     get_committable_i = get_committable_i
+
+    # from pypsa.components.transform
+    rename_component_names = rename_component_names
 
     def __init__(
         self,
@@ -613,142 +622,6 @@ class Components(ComponentsData, ABC):
         """
         return self.dynamic
 
-    def as_dynamic(
-        self,
-        attr: str,
-        snapshots: Sequence | None = None,
-        inds: pd.Index | None = None,
-    ) -> pd.DataFrame:
-        """
-        Get an attribute as a dynamic DataFrame.
-
-        Return a Dataframe for a time-varying component attribute with values for
-        all non-time-varying components filled in with the default values for the
-        attribute.
-
-
-        Parameters
-        ----------
-        n : pypsa.Network
-        component : string
-            Component object name, e.g. 'Generator' or 'Link'
-        attr : string
-            Attribute name
-        snapshots : pandas.Index
-            Restrict to these snapshots rather than n.snapshots.
-        inds : pandas.Index
-            Restrict to these components rather than n.components.index
-
-        Returns
-        -------
-        pandas.DataFrame
-
-        Examples
-        --------
-        >>> import pypsa
-        >>> n = pypsa.examples.ac_dc_meshed()
-        >>> n.components.generators.as_dynamic('p_max_pu', n.snapshots[:2]) # doctest: +SKIP
-        Generator            Manchester Wind  Manchester Gas  Norway Wind  Norway Gas  Frankfurt Wind  Frankfurt Gas
-        snapshot
-        2015-01-01 00:00:00         0.930020             1.0     0.974583         1.0        0.559078            1.0
-        2015-01-01 01:00:00         0.485748             1.0     0.481290         1.0        0.752910            1.0
-
-        """
-        sns = as_index(self.n_save, snapshots, "snapshots")
-        index = self.static.index
-        empty_index = index[:0]  # keep index name and names
-        empty_static = pd.Series([], index=empty_index)
-        static = self.static.get(attr, empty_static)
-        empty_dynamic = pd.DataFrame(index=sns, columns=empty_index)
-        dynamic = self.dynamic.get(attr, empty_dynamic).loc[sns]
-
-        if inds is not None:
-            index = index.intersection(inds)
-
-        diff = index.difference(dynamic.columns)
-        static_to_dynamic = pd.DataFrame({**static[diff]}, index=sns)
-        res = pd.concat([dynamic, static_to_dynamic], axis=1, names=sns.names)[index]
-        res.index.name = sns.name
-        if self.has_scenarios:
-            res.columns.name = "Component"
-            res.columns.names = static.index.names
-        else:
-            res.columns.name = self.name
-        return res
-
-    def as_xarray(
-        self,
-        attr: str,
-        snapshots: Sequence | None = None,
-        inds: pd.Index | None = None,
-    ) -> xarray.DataArray:
-        """
-        Get an attribute as a xarray DataArray.
-
-        Converts component data to a flexible xarray DataArray format, which is
-        particularly useful for optimization routines. The method provides several
-        conveniences:
-
-        1. Supports short attribute name aliases through the `operational_attrs` mapping
-           (e.g., "max_pu" instead of "p_max_pu")
-        2. Automatically handles both static and time-varying attributes
-        3. Creates activity masks with the special "active" attribute name
-        4. Properly handles scenarios if present in the network
-
-        Parameters
-        ----------
-        attr : str
-            Attribute name to retrieve, can be an operational shorthand (e.g., "max_pu")
-            or the full attribute name (e.g., "p_max_pu")
-        snapshots : Sequence | None, optional
-            Snapshots to include. If None, uses all snapshots for time-varying data
-            or returns static data as-is
-        inds : pd.Index | None, optional
-            Component indices to filter by. If None, includes all components
-
-        Returns
-        -------
-        xarray.DataArray
-            The requested attribute data as an xarray DataArray with appropriate dimensions
-
-        Examples
-        --------
-        >>> import pypsa
-        >>> n = pypsa.examples.ac_dc_meshed()
-
-        # Get power output limits for generators for the first two snapshots
-        >>> limit = n.components.generators.as_xarray('p_max_pu', n.snapshots[:2])
-
-        # Use operational attribute shorthand
-        >>> limit = n.components.generators.as_xarray('max_pu', n.snapshots[:2])
-
-        # Get activity mask for lines
-        >>> acitve = n.components.lines.as_xarray('active')
-
-        # Get nominal capacity for specific generators
-        >>> gens = pd.Index(['Manchester Wind', 'Norway Wind'], name='Generator')
-        >>> p_nom = n.components.generators.as_xarray('p_nom', inds=gens)
-
-        """
-        if attr in self.operational_attrs.keys():
-            attr = self.operational_attrs[attr]
-
-        if attr == "active":
-            res = xarray.DataArray(self.get_activity_mask(snapshots, inds))
-        elif attr in self.dynamic.keys() or snapshots is not None:
-            res = xarray.DataArray(self.as_dynamic(attr, snapshots, inds))
-        else:
-            if inds is not None:
-                data = self.static[attr].reindex(inds)
-            else:
-                data = self.static[attr]
-            res = xarray.DataArray(data)
-
-        if self.has_scenarios:
-            # untack the dimension that contains the scenarios
-            res = res.unstack(res.indexes["scenario"].name)
-        return res
-
     @property
     def ds(self) -> xarray.Dataset:
         """Create a xarray data array view of the component."""
@@ -880,68 +753,6 @@ class Components(ComponentsData, ABC):
 
         """
         return self.base_attr + "_nom"
-
-    # TODO move
-    def rename_component_names(self, **kwargs: str) -> None:
-        """
-        Rename component names.
-
-        Rename components and also update all cross-references of the component in
-        the network.
-
-        Parameters
-        ----------
-        **kwargs
-            Mapping of old names to new names.
-
-        Returns
-        -------
-        None
-
-        Examples
-        --------
-        Define some network
-        >>> import pypsa
-        >>> n = pypsa.Network()
-        >>> n.add("Bus", ["bus1"])
-        Index(['bus1'], dtype='object')
-        >>> n.add("Generator", ["gen1"], bus="bus1")
-        Index(['gen1'], dtype='object')
-        >>> c = n.c.buses
-
-        Now rename the bus
-
-        >>> c.rename_component_names(bus1="bus2")
-
-        Which updates the bus components
-
-        >>> c.static.index
-        Index(['bus2'], dtype='object', name='Bus')
-
-        and all references in the network
-
-        >>> n.generators.bus
-        Generator
-        gen1    bus2
-        Name: bus, dtype: object
-
-        """
-        if not all(isinstance(v, str) for v in kwargs.values()):
-            msg = "New names must be strings."
-            raise ValueError(msg)
-
-        # Rename component name definitions
-        self.static = self.static.rename(index=kwargs)
-        for k, v in self.dynamic.items():  # Modify in place
-            self.dynamic[k] = v.rename(columns=kwargs)
-
-        # Rename cross references in network (if attached to one)
-        if self.attached:
-            for c in self.n_save.components.values():
-                col_name = self.name.lower()  # TODO: Generalize
-                cols = [f"{col_name}{port}" for port in c.ports]
-                if cols and not c.static.empty:
-                    c.static[cols] = c.static[cols].replace(kwargs)
 
     @property
     def extendables(self) -> pd.Index:
